moved the specs into the correct place
