+++ src/core/Akka.Tests/IO/UdpIntegrationSpec.cs
@@ -145,89 +145,6 @@ namespace Akka.Tests.IO
         }
 
         [Fact]
-        public void The_UDP_connected_implementation_must_be_able_to_send_several_packet_back_and_forth_with_binding()
-        {
-            var serverAddress = _addresses[0];
-            var clientAddress = _addresses[1];
-            var udpConnection = UdpConnected.Instance.Apply(Sys);
-            var server = CreateTestProbe();
-            udpConnection.Manager.Tell(new UdpConnected.Connect(server, clientAddress, serverAddress), server);
-            server.ExpectMsg<UdpConnected.Connected>();
-            var serverEp = server.LastSender;
-
-            var client = CreateTestProbe();
-            udpConnection.Manager.Tell(new UdpConnected.Connect(client, serverAddress, clientAddress), client);
-            client.ExpectMsg<UdpConnected.Connected>();
-            var clientEp = client.LastSender;
-            var data = ByteString.FromString("Fly little packet!");
-
-            void CheckSendingToClient(int iteration)
-            {
-                serverEp.Tell(UdpConnected.Send.Create(data));
-                client.ExpectMsg<UdpConnected.Received>(x =>
-                {
-                    x.Data.ShouldBe(data);
-                }, hint: $"sending to client failed in {iteration} iteration");
-            }
-
-            void CheckSendingToServer(int iteration)
-            {
-                clientEp.Tell(UdpConnected.Send.Create(data));
-                server.ExpectMsg<UdpConnected.Received>(x =>
-                {
-                    x.Data.ShouldBe(data);
-
-                }, hint: $"sending to client failed in {iteration} iteration");
-            }
-
-            const int iterations = 20;
-            for (int i = 1; i <= iterations; i++) CheckSendingToServer(i);
-            for (int i = 1; i <= iterations; i++) CheckSendingToClient(i);
-            for (int i = 1; i <= iterations; i++)
-            {
-                if (i % 2 == 0) CheckSendingToServer(i);
-                else CheckSendingToClient(i);
-            }
-        }
-
-        [Fact]
-        public void The_UDP_connected_implementation_must_to_send_batch_writes_and_reads()
-        {
-            var serverAddress = _addresses[0];
-            var clientAddress = _addresses[1];
-            var udpConnection = UdpConnected.Instance.Apply(Sys);
-            var server = CreateTestProbe();
-            udpConnection.Manager.Tell(new UdpConnected.Connect(server, clientAddress, serverAddress), server);
-            server.ExpectMsg<UdpConnected.Connected>();
-            var serverEp = server.LastSender;
-
-            var client = CreateTestProbe();
-            udpConnection.Manager.Tell(new UdpConnected.Connect(client, serverAddress, clientAddress), client);
-            client.ExpectMsg<UdpConnected.Connected>();
-            var clientEp = client.LastSender;
-            var data = ByteString.FromString("Fly little packet!");
-
-            // queue 3 writes
-            clientEp.Tell(UdpConnected.Send.Create(data));
-            clientEp.Tell(UdpConnected.Send.Create(data));
-            clientEp.Tell(UdpConnected.Send.Create(data));
-
-            var raw = server.ReceiveN(3);
-            var msgs = raw.Cast<UdpConnected.Received>();
-            msgs.Sum(x => x.Data.Count).Should().Be(data.Count * 3);
-            server.ExpectNoMsg(100.Milliseconds());
-
-            // repeat in the other direction
-            serverEp.Tell(UdpConnected.Send.Create(data));
-            serverEp.Tell(UdpConnected.Send.Create(data));
-            serverEp.Tell(UdpConnected.Send.Create(data));
-
-            raw = client.ReceiveN(3);
-            msgs = raw.Cast<UdpConnected.Received>();
-            msgs.Sum(x => x.Data.Count).Should().Be(data.Count * 3);
-        }
-
-        [Fact]
         public void The_UDP_Fire_and_Forget_implementation_must_call_SocketOption_beforeBind_method_before_bind()
         {
             var commander = CreateTestProbe();
